feat : whisper transcibe audio endpoint
This commit consists of the endpoint which redirects to openAI
transcribe audio endpoint and uses whisper as the conversion model.
Along with the endpoint, I have also added the example and reference use
case of how this can be called. It can be found inside the `whisper.tsx`
file and to just test what is does load that component into index.tsx.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>