fix : Testing and fixing the whisper API
Thanks to Shwet's OpenAI secret key using which I was actually able to
interact with the output and run it. With this one challange is that
API needs a ReadStream for the audio file. But, Currently, I have
hardcoded the file from my machine. It needs to be made generic for the
uploaded file service. I have tried my hands with multer but it has
lots of NextJS and Express compatibility issues. Need to see how this
can be achieved.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>