feat: replaced whisper with Web Speech API
I found Web Speech API a good alternative of Whisper Model which is a
speech to text model by OpenAI. Reason why I would choose to do so are
simple as we don't want a complicated layer of protocol which stores
voice recordings (maybe). This client side utility would make us data
lovers less possessive. We may have less accuracy as this is not an
modern world's AI model but it works upto the mark.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>