feat : Virtual Tourism landing description
I have been developing the Maps and the Model work since a while and I
am insomniac so decided to take up this golden opportunity to do the
most of the nits off the work. Up next in the row I have series of
features in mind. Speech to Text and integrating ChatGPT which will give
me beatiful prompts discussing about the cities and what are the best
things to do over there.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>